25.01

+++ PortalPracy/php/delete_reference.php
@@ -1,15 +1,16 @@
 <?php
+if(isset($_SESSION['name'])) {
+    header("Cache-Control: no-cache, must-revalidate");
+    header("Expires: Mon, 26 Jul 1997 05:00:00 GMT");
+    error_reporting(0);
 
-header("Cache-Control: no-cache, must-revalidate");
-header("Expires: Mon, 26 Jul 1997 05:00:00 GMT");
-error_reporting(0);
+    if (!isset($_GET['id'])) { 
+        echo 'false';
+        exit;
+    }
 
-if (!isset($_GET['id'])) { 
-    echo 'false';
-    exit;
+    $id =$_GET['id'];
+    include_once 'classDatabase.php';
+    $db = new Database();
+    $db->delete_reference($id);
 }
-
-$id =$_GET['id'];
-include_once 'classDatabase.php';
-$db = new Database();
-$db->delete_reference($id);
